Added 'auth_key/auth_secret' credentials handling. Added 'requester' class.
Credentials are now required by MtGox in order to access the API.
You will need to open an account with MtGox and obtain your credentials (key and secret) from the security section in your MtGox settings.  You do not need to be verified for trading to qualify for these credentials.